+++ sls-20201230/core/Models/IndexJsonKey.cs
@@ -8,22 +8,26 @@ using Tea;
 
 namespace AlibabaCloud.SDK.Sls20201230.Models
 {
-    public class IndexKeysValue : TeaModel {
-        [NameInMap("chn")]
+    public class IndexJsonKey : TeaModel {
+        [NameInMap("alias")]
         [Validation(Required=false)]
-        public bool? Chn { get; set; }
+        public string Alias { get; set; }
 
         [NameInMap("caseSensitive")]
         [Validation(Required=false)]
         public bool? CaseSensitive { get; set; }
 
+        [NameInMap("chn")]
+        [Validation(Required=false)]
+        public bool? Chn { get; set; }
+
+        [NameInMap("doc_value")]
+        [Validation(Required=false)]
+        public bool? DocValue { get; set; }
+
         [NameInMap("token")]
         [Validation(Required=false)]
         public List<string> Token { get; set; }
-
-        [NameInMap("alias")]
-        [Validation(Required=false)]
-        public string Alias { get; set; }
 
         /// <summary>
         /// This parameter is required.
@@ -31,10 +35,6 @@ namespace AlibabaCloud.SDK.Sls20201230.Models
         [NameInMap("type")]
         [Validation(Required=false)]
         public string Type { get; set; }
-
-        [NameInMap("doc_value")]
-        [Validation(Required=false)]
-        public bool? DocValue { get; set; }
 
     }